PC/M: fix FDD selection handling at port 0C2h
Bit 4..0 behind port 0C2h are used as an direct selection of
Floppy Disk Drives. There is no hardware decoder that will
decode a BCD number into a singel selction line. Bit 0 selects
the first possible drive (/SE0 and /LCK0), Bit 1 the second
one (/SEL1 and /LCK1) and so on.

The emulator must interpret the lower four bits as an 1-of-4
code, not as a number. Further the emulator have to respect the
special drive numbering in PC/M BIOS:

  drive A --> BDOS FCB 0 --> RAM floppy, will never reach this code
  drive B --> BDOS FCB 1 --> 1st Floppy Disk Drive
  drive C --> BDOS FCB 2 --> 2nd Floppy Disk Drive
  drive D --> BDOS FCB 3 --> 3rd Floppy Disk Drive
  drive E --> BDOS FCB 4 --> 4th Floppy Disk Drive, but this is not
                             possible for PC/M BIOS

The corresponding system code in PC/M BIOS can found at the
assembler mark 'drrdy' and 'drr1' in PCMDISK.MAC:

drrdy:
	push	bc
	ld	a,(ix+1)
	and	3
	inc	a
	ld	b,a
	ld	a,80h
drr1:	rlca
	djnz	drr1
	ld	b,a
	...

Register IX points to the FCB in BDOS. The drive number (IX+1)
will be loaded to register A. The loop on mark 'drr1' create a
bit mask corrosponding to the value in register A (drive number).

Signed-off-by: Stephan Linz <linz@li-pro.net>